requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
Django==1.10.5
djangorestframework==3.5.4
packaging==16.8
PyExecJS==1.4.0
pyjsparser==2.4.5
pyparsing==2.1.10
pytz==2016.10
requests==2.13.0
six==1.10.0
tzlocal==1.3
webencodings==0.5
